Test now compiles and completes, but doesn't quite work right yet.
One key bugfix was moving from an array of q_entry_t to a vector
of q_entry_t for the circular buffer in wire_t, so that push_back(wire_t)
works correctly.  push_back() relies on your type having a working
operator=, and that means that you need to overload the operator
yourself if you need to do anything special (like deep-copying arrays).
I didn't feel like doing that just yet, so I switched to a vector
which does it for me.  Can revisit this later.

Also turned off DEBUG in most places.